backend/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
Werkzeug==3.0.3
torch==2.3.0
torchvision==0.18.0
opencv-python==4.9.0.80
numpy==1.26.4
Pillow==10.3.0
scipy
PyWavelets
